# docs/OPENAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Dhruva API",
    "description": "Backend API for communicating with the Dhruva platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/services/details/list_services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": " List Services",
        "operationId": "_list_services_services_details_list_services_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response  List Services Services Details List Services Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceListResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/details/view_service": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " View Service Details",
        "operationId": "_view_service_details_services_details_view_service_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceViewResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/details/list_models": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": " List Models",
        "operationId": "_list_models_services_details_list_models_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response  List Models Services Details List Models Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelViewResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/details/view_model": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " View Model Details",
        "operationId": "_view_model_details_services_details_view_model_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelViewResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseErrorResponse": {
        "title": "BaseErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/exception__base_error___ErrorDetail"
          }
        }
      },
      "ClientErrorResponse": {
        "title": "ClientErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/exception__client_error___ErrorDetail"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Model": {
        "title": "Model",
        "required": [
          "modelId",
          "version",
          "submittedOn",
          "updatedOn",
          "name",
          "description",
          "refUrl",
          "task",
          "languages",
          "license",
          "domain",
          "inferenceEndPoint",
          "submitter"
        ],
        "type": "object",
        "properties": {
          "modelId": {
            "title": "Modelid",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "submittedOn": {
            "title": "Submittedon",
            "type": "integer"
          },
          "updatedOn": {
            "title": "Updatedon",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "refUrl": {
            "title": "Refurl",
            "type": "string"
          },
          "task": {
            "$ref": "#/components/schemas/_ULCATask"
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inferenceEndPoint": {
            "$ref": "#/components/schemas/_InferenceEndPoint"
          },
          "benchmarks": {
            "title": "Benchmarks",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/schema__services__common__dhruva_model___Benchmark"
            }
          },
          "submitter": {
            "$ref": "#/components/schemas/_Submitter"
          }
        }
      },
...
      "ModelViewRequest": {
        "title": "ModelViewRequest",
        "required": [
          "modelId"
        ],
        "type": "object",
        "properties": {
          "modelId": {
            "title": "Modelid",
            "type": "string"
          }
        }
      },
      "ModelViewResponse": {
        "title": "ModelViewResponse",
        "required": [
          "modelId",
          "name",
          "description",
          "languages",
          "domain",
          "submitter",
          "license",
          "inferenceEndPoint",
          "source",
          "task"
        ],
        "type": "object",
        "properties": {
          "modelId": {
            "title": "Modelid",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "submitter": {
            "title": "Submitter",
            "type": "string"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "inferenceEndPoint": {
            "title": "Inferenceendpoint",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        }
      },
...
      "ServiceListResponse": {
        "title": "ServiceListResponse",
        "required": [
          "serviceId",
          "name",
          "serviceDescription",
          "hardwareDescription",
          "publishedOn",
          "modelId",
          "task",
          "languages"
        ],
        "type": "object",
        "properties": {
          "serviceId": {
            "title": "Serviceid",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "serviceDescription": {
            "title": "Servicedescription",
            "type": "string"
          },
          "hardwareDescription": {
            "title": "Hardwaredescription",
            "type": "string"
          },
          "publishedOn": {
            "title": "Publishedon",
            "type": "integer"
          },
          "modelId": {
            "title": "Modelid",
            "type": "string"
          },
          "healthStatus": {
            "$ref": "#/components/schemas/ServiceStatus"
          },
          "benchmarks": {
            "title": "Benchmarks",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/schema__services__response__service_response___Benchmark"
              }
            }
          },
          "task": {
            "$ref": "#/components/schemas/_ULCATask"
          },
          "languages": {
            "title": "Languages",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ServiceStatus": {
        "title": "ServiceStatus",
        "required": [
          "status",
          "lastUpdated"
        ],
        "type": "object",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "lastUpdated": {
            "title": "Lastupdated",
            "type": "string"
          }
        }
      },
...
      "ServiceViewRequest": {
        "title": "ServiceViewRequest",
        "required": [
          "serviceId"
        ],
        "type": "object",
        "properties": {
          "serviceId": {
            "title": "Serviceid",
            "type": "string"
          }
        }
      },
      "ServiceViewResponse": {
        "title": "ServiceViewResponse",
        "required": [
          "serviceId",
          "name",
          "serviceDescription",
          "hardwareDescription",
          "publishedOn",
          "modelId",
          "model"
        ],
        "type": "object",
        "properties": {
          "serviceId": {
            "title": "Serviceid",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "serviceDescription": {
            "title": "Servicedescription",
            "type": "string"
          },
          "hardwareDescription": {
            "title": "Hardwaredescription",
            "type": "string"
          },
          "publishedOn": {
            "title": "Publishedon",
            "type": "integer"
          },
          "modelId": {
            "title": "Modelid",
            "type": "string"
          },
          "healthStatus": {
            "$ref": "#/components/schemas/ServiceStatus"
          },
          "benchmarks": {
            "title": "Benchmarks",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/schema__services__response__service_response___Benchmark"
              }
            }
          },
          "model": {
            "$ref": "#/components/schemas/Model"
          },
          "key_usage": {
            "title": "Key Usage",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_ApiKey"
            },
            "default": []
          },
          "total_usage": {
            "title": "Total Usage",
            "type": "integer",
            "default": 0
          }
        }
      },
...
      "TokenType": {
        "title": "TokenType",
        "enum": [
          "AUTH_TOKEN",
          "API_KEY"
        ],
        "description": "An enumeration."
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
      "_ApiKey": {
        "title": "_ApiKey",
        "required": [
          "name",
          "masked_key",
          "active",
          "type",
          "created_timestamp",
          "services",
          "data_tracking",
          "id"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "masked_key": {
            "title": "Masked Key",
            "type": "string"
          },
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "created_timestamp": {
            "title": "Created Timestamp",
            "type": "string",
            "format": "date-time"
          },
          "services": {
            "title": "Services",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_ServiceUsage"
            }
          },
          "data_tracking": {
            "title": "Data Tracking",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        }
      },
      "_BenchmarkMetric": {
        "title": "_BenchmarkMetric",
        "required": [
          "metricName",
          "score"
        ],
        "type": "object",
        "properties": {
          "metricName": {
            "title": "Metricname",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "string"
          }
        }
      },
...
      "_InferenceEndPoint": {
        "title": "_InferenceEndPoint",
        "required": [
          "schema"
        ],
        "type": "object",
        "properties": {
          "schema": {
            "$ref": "#/components/schemas/_Schema"
          }
        }
      },
      "_LanguagePair": {
        "title": "_LanguagePair",
        "required": [
          "sourceLanguage"
        ],
        "type": "object",
        "properties": {
          "sourceLanguage": {
            "title": "Sourcelanguage",
            "type": "string"
          },
          "targetLanguage": {
            "title": "Targetlanguage",
            "type": "string"
          }
        }
      },
      "_ModelProcessingType": {
        "title": "_ModelProcessingType",
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      },
...
      "_OAuthId": {
        "title": "_OAuthId",
        "required": [
          "oauthId",
          "provider"
        ],
        "type": "object",
        "properties": {
          "oauthId": {
            "title": "Oauthid",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        }
      },
      "_Schema": {
        "title": "_Schema",
        "required": [
          "request",
          "response"
        ],
        "type": "object",
        "properties": {
          "modelProcessingType": {
            "$ref": "#/components/schemas/_ModelProcessingType"
          },
          "request": {
            "title": "Request",
            "type": "object"
          },
          "response": {
            "title": "Response",
            "type": "object"
          }
        }
      },
      "_ServiceUsage": {
        "title": "_ServiceUsage",
        "required": [
          "service_id",
          "usage"
        ],
        "type": "object",
        "properties": {
          "service_id": {
            "title": "Service Id",
            "type": "string"
          },
          "usage": {
            "title": "Usage",
            "type": "integer"
          }
        }
      },
      "_Submitter": {
        "title": "_Submitter",
        "required": [
          "name",
          "aboutMe",
          "team"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "aboutMe": {
            "title": "Aboutme",
            "type": "string"
          },
          "team": {
            "title": "Team",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_SubmitterDetails"
            }
          }
        }
      },
      "_SubmitterDetails": {
        "title": "_SubmitterDetails",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "aboutMe": {
            "title": "Aboutme",
            "type": "string"
          },
          "oauthId": {
            "$ref": "#/components/schemas/_OAuthId"
          }
        }
      },
...
      "_ULCATask": {
        "title": "_ULCATask",
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/_ULCATaskType"
          }
        }
      },
      "_ULCATaskType": {
        "title": "_ULCATaskType",
        "enum": [
          "asr",
          "translation",
          "tts",
          "transliteration",
          "ner",
          "sts",
          "vad"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "exception__base_error___ErrorDetail": {
        "title": "_ErrorDetail",
        "required": [
          "kind",
          "message"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "exception__client_error___ErrorDetail": {
        "title": "_ErrorDetail",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "schema__services__common__dhruva_model___Benchmark": {
        "title": "_Benchmark",
        "required": [
          "name",
          "description",
          "domain",
          "score"
        ],
        "type": "object",
        "properties": {
          "benchmarkId": {
            "title": "Benchmarkid",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "createdOn": {
            "title": "Createdon",
            "type": "string",
            "format": "date-time"
          },
          "languages": {
            "$ref": "#/components/schemas/_LanguagePair"
          },
          "score": {
            "title": "Score",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_BenchmarkMetric"
            }
          }
        }
      },
...
      "schema__services__response__service_response___Benchmark": {
        "title": "_Benchmark",
        "required": [
          "output_length",
          "generated",
          "actual",
          "throughput",
          "50%",
          "99%",
          "language"
        ],
        "type": "object",
        "properties": {
          "output_length": {
            "title": "Output Length",
            "type": "integer"
          },
          "generated": {
            "title": "Generated",
            "type": "integer"
          },
          "actual": {
            "title": "Actual",
            "type": "integer"
          },
          "throughput": {
            "title": "Throughput",
            "type": "integer"
          },
          "50%": {
            "title": "50%",
            "type": "number"
          },
          "99%": {
            "title": "99%",
            "type": "number"
          },
          "language": {
            "title": "Language",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}